JS innerHTML, array/loop JS, weekdays
removed row/column HTML and placed in JS as innerHTML function/set css to col-2 to space evenly across row.

created an array for the days of the week so we could loop through them with API later.

set an array variable for the days of the week.

created a forEach function.

went into innerHTML on JS and assigned template literals to change said day.