Forecast API integration / console log variable
done for -
icon
max temp
min temp
time

used the console log for (response.data.daily) and made that into the weeklyForecast variable output.

used that variable to correspond to the forEach function.

changed parameter inside the forEach function to reflect the new idea for the data we are now putting in. No longer need the word day as it is much more than just a day now.